+++ nitrogen/lib/nitrogen.dart
@@ -2,6 +2,10 @@ import 'dart:async';
 
 import 'package:build/build.dart';
 import 'package:glob/glob.dart';
+import 'package:path/path.dart';
+import 'package:yaml/yaml.dart';
+
+import 'package:nitrogen/src/configuration/build_configuration.dart';
 import 'package:nitrogen/src/configuration/configuration.dart';
 import 'package:nitrogen/src/file_system.dart';
 import 'package:nitrogen/src/generators/asset_generator.dart';
@@ -9,25 +13,34 @@ import 'package:nitrogen/src/generators/theme_generator.dart';
 import 'package:nitrogen/src/lints/reserved_keyword_lint.dart';
 import 'package:nitrogen/src/nitrogen_exception.dart';
 import 'package:nitrogen/src/walker.dart';
-import 'package:path/path.dart';
-import 'package:yaml/yaml.dart';
 
 /// Creates a [NitrogenBuilder].
-Builder nitrogenBuilder(BuilderOptions options) => NitrogenBuilder();
+Builder nitrogenBuilder(BuilderOptions options) {
+  try {
+    BuildConfiguration.lint(options.config);
+    return NitrogenBuilder(BuildConfiguration.parse(options.config));
+
+  } on NitrogenException {
+    return EmptyBuilder();
+  }
+}
 
 /// A Nitrogen builder.
 class NitrogenBuilder extends Builder {
 
   static final _pubspec = Glob('pubspec.yaml');
   static final _assets = Glob('assets/**');
+
+  final BuildConfiguration _configuration;
+
+  /// Creates a [NitrogenBuilder].
+  NitrogenBuilder(this._configuration);
   
   @override
   FutureOr<void> build(BuildStep buildStep) async {
     try {
       final pubspec = loadYaml(await buildStep.readAsString(await buildStep.findAssets(_pubspec).first));
-
-      Configuration.lint(pubspec);
-      final configuration = Configuration.parse(pubspec);
+      final configuration = Configuration.merge(_configuration, pubspec);
 
       final walker = Walker(configuration.package, configuration.flutterAssets, configuration.key.call);
 
@@ -38,8 +51,8 @@ class NitrogenBuilder extends Builder {
 
       lintReservedKeyword(assets);
 
-      final assetsOutput = AssetId(buildStep.inputId.package, join('lib', 'src', 'assets.nitrogen.dart'));
-      if (configuration.fallbackTheme case final fallback?) {
+      final assetsOutput = AssetId(buildStep.inputId.package, configuration.assets.output);
+      if (configuration.themes case (:final fallback, :final output)) {
         var themes = assets;
         var fallbackTheme = assets;
         for (final segment in split(fallback).skip(1)) {
@@ -53,7 +66,7 @@ class NitrogenBuilder extends Builder {
         );
 
         await buildStep.writeAsString(
-          AssetId(buildStep.inputId.package, join('lib', 'src', 'asset_themes.nitrogen.dart')),
+          AssetId(buildStep.inputId.package, output),
           ThemeGenerator(configuration.prefix, themes, fallbackTheme).generate(),
         );
 
@@ -72,9 +85,19 @@ class NitrogenBuilder extends Builder {
   @override
   Map<String, List<String>> get buildExtensions => {
     r'$package$': [
-      'lib/src/assets.nitrogen.dart',
-      'lib/src/asset_themes.nitrogen.dart',
+      _configuration.assets.output,
+      if (_configuration.themes != null)
+        _configuration.themes!.output,
     ],
   };
   
+}
+
+/// A stub builder for when Nitrogen fails to build.
+class EmptyBuilder extends Builder {
+  @override
+  void build(BuildStep buildStep) {}
+
+  @override
+  Map<String, List<String>> get buildExtensions => {};
 }