+++ nitrogen/lib/src/configuration/configuration.dart
@@ -1,46 +1,27 @@
 import 'package:build/build.dart';
 import 'package:meta/meta.dart';
-import 'package:nitrogen/src/nitrogen_exception.dart';
-import 'package:nitrogen/src/configuration/key.dart';
 import 'package:yaml/yaml.dart';
+
+import 'package:nitrogen/src/configuration/build_configuration.dart';
+import 'package:nitrogen/src/nitrogen_exception.dart';
 
 /// Nitrogen's configuration.
 final class Configuration {
 
-  /// The Nitrogen section's valid keys.
-  static const keys = { 'package', 'prefix', 'key', 'themes' };
-
-  /// Lints the pubspec.
-  static void lint(YamlMap? pubspec) {
-    final nitrogen = pubspec?.nodes['nitrogen'].as<YamlMap>();
-
-    for (final key in {...?nitrogen?.keys }..removeAll(keys)) {
-      log.warning(nitrogen?.nodes[key]!.span.message('Unknown key, "$key", in pubspec.yaml\'s nitrogen section. See https://github.com/forus-labs/cauldron/tree/master/nitrogen#configuration for valid configuration options.'));
-    }
-
-    final themes = nitrogen?.nodes['themes'].as<YamlMap>();
-    for (final key in {...?themes?.keys }..removeAll({ 'fallback' })) {
-      log.warning(themes?.nodes[key]!.span.message('Unknown key, "$key", in pubspec.yaml\'s nitrogen section. See https://github.com/forus-labs/cauldron/tree/master/nitrogen#themes for valid configuration options.'));
-    }
-  }
-
-
   /// Parses the configuration from the project's pubspec.yaml.
-  factory Configuration.parse(YamlMap pubspec) {
-    final nitrogen = pubspec.nodes['nitrogen'].as<YamlMap>();
-    return Configuration(
-      package: parsePackage(pubspec.nodes['name'], nitrogen?.nodes['package']),
-      prefix: parsePrefix(nitrogen?.nodes['prefix']),
-      key: Key.parse(nitrogen?.nodes['key']),
-      fallbackTheme: parseFallbackTheme(nitrogen?.nodes['themes']),
-      flutterAssets: parseFlutterAssets(pubspec.nodes['flutter'].as<YamlMap>()?.nodes['assets']),
-    );
-  }
+  factory Configuration.merge(BuildConfiguration configuration, YamlMap pubspec) => Configuration(
+    package: parsePackage(pubspec.nodes['name'], enabled: configuration.package),
+    prefix: configuration.prefix,
+    key: configuration.key,
+    assets: configuration.assets,
+    themes: configuration.themes,
+    flutterAssets: parseFlutterAssets(pubspec.nodes['flutter'].as<YamlMap>()?.nodes['assets']),
+  );
 
   /// Parses the package name from the project's pubspec.yaml.
   @visibleForTesting
-  static String? parsePackage(YamlNode? name, YamlNode? enabled) {
-    switch ((name?.value, enabled?.value)) {
+  static String? parsePackage(YamlNode? name, {required bool enabled}) {
+    switch ((name?.value, enabled)) {
       case (final String name, true):
         return name;
 
@@ -50,39 +31,6 @@ final class Configuration {
 
       case (_, bool? _):
         return null;
-
-      default:
-        log.severe(enabled?.span.message('Unable to read package name. See https://github.com/forus-labs/cauldron/tree/master/nitrogen#package.'));
-        throw NitrogenException();
-    }
-  }
-
-  /// Parses the class prefix from the nitrogen section in the project's pubspec.yaml.
-  @visibleForTesting
-  static String parsePrefix(YamlNode? node) {
-    switch (node?.value) {
-      case final String? prefix:
-        return prefix ?? '';
-
-      default:
-        log.severe(node!.span.message('Unable to read prefix. See https://github.com/forus-labs/cauldron/tree/master/nitrogen#prefix.'));
-        throw NitrogenException();
-    }
-  }
-
-  /// Parses the path to the fallback theme.
-  @visibleForTesting
-  static String? parseFallbackTheme(YamlNode? node) {
-    switch (node?.value) {
-      case null:
-        return null;
-
-      case { 'fallback': final String fallback }:
-        return fallback;
-
-      case _:
-        log.severe(node!.span.message('Unable to read themes. See https://github.com/forus-labs/cauldron/tree/master/nitrogen#themes.'));
-        throw NitrogenException();
     }
   }
 
@@ -111,8 +59,11 @@ final class Configuration {
   /// The function for generating asset keys.
   final String Function(List<String>) key;
 
-  /// The path to the fallback theme.
-  final String? fallbackTheme;
+  /// The output location of the generated assets.
+  final ({String output,}) assets;
+
+  /// The fallback theme and output location of the generated themes.
+  final ({String fallback, String output})? themes;
 
   /// The flutter assets.
   final Set<String> flutterAssets;
@@ -122,7 +73,8 @@ final class Configuration {
     required this.package,
     required this.prefix,
     required this.key,
-    required this.fallbackTheme,
+    required this.assets,
+    required this.themes,
     required this.flutterAssets,
   });