+++ nitrogen/lib/src/lints/reserved_keyword_lint.dart
@@ -1,4 +1,5 @@
 import 'package:build/build.dart';
+
 import 'package:nitrogen/src/file_system.dart';
 import 'package:nitrogen/src/nitrogen_exception.dart';