+++ nitrogen/lib/src/walker.dart
@@ -1,8 +1,9 @@
 import 'dart:collection';
 
 import 'package:build/build.dart';
-import 'package:nitrogen/src/file_system.dart';
 import 'package:nitrogen_types/nitrogen_types.dart';
+
+import 'package:nitrogen/src/file_system.dart';
 
 /// A walker that recursively walks through a directory and converts all entities into [Entity]s.
 final class Walker {