+++ nitrogen/test/src/file_system_test.dart
@@ -1,6 +1,7 @@
-import 'package:nitrogen/src/file_system.dart';
 import 'package:nitrogen_types/nitrogen_types.dart';
 import 'package:test/test.dart';
+
+import 'package:nitrogen/src/file_system.dart';
 
 void main() {
   group('AssetDirectory', () {