+++ nitrogen/test/src/generators/theme_generator_test.dart
@@ -1,8 +1,9 @@
 import 'package:code_builder/code_builder.dart';
-import 'package:nitrogen/src/file_system.dart';
-import 'package:nitrogen/src/generators/theme_generator.dart';
 import 'package:nitrogen_types/nitrogen_types.dart';
 import 'package:test/test.dart';
+
+import 'package:nitrogen/src/file_system.dart';
+import 'package:nitrogen/src/generators/theme_generator.dart';
 
 const _classes = r'''
 import 'package:nitrogen_types/nitrogen_types.dart';