+++ nitrogen/test/src/lints/reserved_keyword_lint_test.dart
@@ -1,10 +1,11 @@
 import 'package:build/build.dart';
 import 'package:build_test/build_test.dart';
+import 'package:nitrogen_types/nitrogen_types.dart';
+import 'package:test/test.dart';
+
 import 'package:nitrogen/src/file_system.dart';
 import 'package:nitrogen/src/lints/reserved_keyword_lint.dart';
 import 'package:nitrogen/src/nitrogen_exception.dart';
-import 'package:nitrogen_types/nitrogen_types.dart';
-import 'package:test/test.dart';
 
 void main() {
   group('lintReservedKeyword', () {