+++ nitrogen/test/src/walker_test.dart
@@ -1,8 +1,9 @@
 import 'package:build/build.dart';
-import 'package:nitrogen/src/file_system.dart';
-import 'package:nitrogen/src/walker.dart';
 import 'package:nitrogen_types/nitrogen_types.dart';
 import 'package:test/test.dart';
+
+import 'package:nitrogen/src/file_system.dart';
+import 'package:nitrogen/src/walker.dart';
 
 void main() {
   group('walk(...)', () {