+++ nitrogen/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0
+* Add configuration for output locations.
+* Move configuration from `pubspec.yaml` to `build.yaml`.
+* Remove generation of `flutter_extension`.
+
 ## 0.2.0
 * Add `flutter-extension` option.
 * Change `file` asset key option to `file-name`.